requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.2
frozenlist==1.7.0
future==1.0.0
git-filter-repo==2.47.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
line-bot-sdk==3.19.0
MarkupSafe==3.0.2
multidict==6.6.4
oauthlib==3.3.1
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.20.1
